Refactor CinemaHall component and extract styles
Moved CinemaHall component to its own directory, extracted styles into a separate CinemaHall.styles.ts file, and updated imports accordingly. Added an index.ts for cleaner exports.
