+++ src/screens/Watch/WatchReserveSeatScreen/components/CinemaHall/index.ts
@@ -1,0 +1,1 @@
+export { CinemaHall } from './CinemaHall';
